refact : jwt validator 추가해서 jwt 검증하고 refactoring
