+++ src/main/java/kuit3/backend/common/argument_resolver/JwtAuthHandlerArgumentResolver.java
@@ -1,14 +1,10 @@
 package kuit3.backend.common.argument_resolver;
 
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jws;
-import io.jsonwebtoken.Jwts;
 import jakarta.servlet.http.HttpServletRequest;
-import kuit3.backend.common.interceptor.JwtAuthInterceptor;
+import kuit3.backend.jwt.JwtValidator;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.MethodParameter;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
@@ -19,8 +15,11 @@ import org.springframework.web.method.support.ModelAndViewContainer;
 @Component
 public class JwtAuthHandlerArgumentResolver implements HandlerMethodArgumentResolver {
 
-    @Value("${JWT_SECRET_KEY}")
-    private String JWT_SECRET_KEY;
+    private final JwtValidator jwtValidator;
+
+    public JwtAuthHandlerArgumentResolver(JwtValidator jwtValidator) {
+        this.jwtValidator = jwtValidator;
+    }
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
@@ -34,15 +33,7 @@ public class JwtAuthHandlerArgumentResolver implements HandlerMethodArgumentReso
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
         HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
         String jwt = request.getHeader("Authorization");
-        if (jwt != null && jwt.startsWith("Bearer ")) {
-            String token = jwt.substring(7); // "Bearer " 제거
-            return Jwts.parserBuilder()
-                    .setSigningKey(JWT_SECRET_KEY).build()
-                    .parseClaimsJws(token)
-                    .getBody().get("userId");
-        } else {
-            // JWT 토큰이 올바르게 전달되지 않은 경우에 대한 처리
-            throw new IllegalArgumentException("Invalid JWT token");
-        }
+        jwtValidator.isExpiredToken(jwt.substring(7));
+        return jwtValidator.getLongValueByTokenPayload(jwt, "userId");
     }
 }
