+++ src/main/java/kuit3/backend/jwt/JwtProvider.java
@@ -40,28 +40,6 @@ public class JwtProvider {
                 .compact();
     }
 
-    public boolean isExpiredToken(String token) throws JwtInvalidTokenException {
-        try {
-            Jws<Claims> claims = Jwts.parserBuilder()
-                    .setSigningKey(JWT_SECRET_KEY).build()
-                    .parseClaimsJws(token);
-            return claims.getBody().getExpiration().before(new Date());
-
-        } catch (ExpiredJwtException e) {
-            return true;
-
-        } catch (UnsupportedJwtException e) {
-            throw new JwtUnsupportedTokenException(UNSUPPORTED_TOKEN_TYPE);
-        } catch (MalformedJwtException e) {
-            throw new JwtMalformedTokenException(MALFORMED_TOKEN);
-        } catch (IllegalArgumentException e) {
-            throw new JwtInvalidTokenException(INVALID_TOKEN);
-        } catch (JwtException e) {
-            log.error("[JwtTokenProvider.validateAccessToken]", e);
-            throw e;
-        }
-    }
-
     public String getPrincipal(String token) {
         return Jwts.parserBuilder()
                 .setSigningKey(JWT_SECRET_KEY).build()
